alter: Integer to Long

+++ src/main/java/com/example/authentication/model/User.java
@@ -30,7 +30,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer userId;
+    private Long userId;
 
     @Column(unique = true, nullable = false)
     private String username;
@@ -61,7 +61,7 @@ public class User {
 
     @OneToMany(mappedBy = "user")
     private List<LoginAudit> loginAudits;
-    
+
     public User(String username, String phoneNumber, String email, Type type) {
         this.username = username;
         this.phoneNumber = phoneNumber;
